Fixed stack overflow (crash on arm64 but not x86)
The stack overflow bug exists on all platfoms (there is a 33 byte memcpy into a 32 byte buffer on the stack), but for whatever reason, it doesn’t cause issues on x86/x86_64, but crashes on arm64. Simple fix to just increase the receiving buffer to be the correct size, and now it works on all platforms without crashing.